Prevents you from deleting yourself.

+++ bonfire/application/core_modules/users/controllers/settings.php
@@ -181,7 +181,7 @@ class Settings extends Admin_Controller {
 			$this->auth->restrict('Bonfire.Users.Manage');
 		
 			$user = $this->user_model->find($id);
-			if (isset($user) && has_permission('Permissions.'.$user->role_name.'.Manage'))
+			if (isset($user) && has_permission('Permissions.'.$user->role_name.'.Manage') && $user->id != $id)
 			{
 				if ($this->user_model->delete($id))
 				{
@@ -196,7 +196,14 @@ class Settings extends Admin_Controller {
 			}
 			else
 			{
-				Template::set_message(sprintf(lang('us_unauthorized'),$user->role_name), 'error');
+				if ($user->id == $id)
+				{
+					Template::set_message(lang('us_self_delete'), 'error');
+				}
+				else
+				{
+					Template::set_message(sprintf(lang('us_unauthorized'),$user->role_name), 'error');	
+				}				
 			}
 		}
 		else
